+++ axelor-production/src/main/java/com/axelor/apps/production/translation/ITranslation.java
@@ -22,4 +22,6 @@ public interface ITranslation {
   public static final String MANUFACTURING_APP_NAME = /*$$(*/ "value:Manufacturing"; /*)*/
   public static final String WORK_IN_PROGRESS_VALUATION = /*$$(*/
       "Work in progress valuation"; /*)*/
+  public static final String DETAILED_WORK_IN_PROGRESS_VALUATION = /*$$(*/
+      "Detailed work in progress valuation"; /*)*/
 }